Add the index.js file that we started with
This file can be used to serve the app on a webpage using localhost.
Once we have website infrastructure up and running this is what we would
probably use to serve the website.

Currently, the images don't load with this approach so need to figure why
that happens. For the time being we can continue developing using the
.html file.